Chriswf

+++ server/index.js
@@ -27,6 +27,9 @@ var app = require('express')(),
     saveUninitialized: true
   });
 
+app.use(session) // ATTACH SESSION
+// console.log("initial", session) //Session exists at this point
+
 //Auth0
 
 app.use(passport.initialize());
@@ -74,10 +77,10 @@ passport.deserializeUser(function(obj, done) {
   done(null, obj);
 });
 
-let you;
-app.get('/login', passport.authenticate('auth0'), function(req, res, next) {
-  you = req.user;
-  session.user = req.user;
+let you
+app.get("/login", passport.authenticate("auth0"), function(req, res, next) {
+  you = req.user
+  req.session.user = req.user
   req.user.rank === 3
     ? res.redirect('http://localhost:3000/student')
     : res.redirect('http://localhost:3000/mentorview');
@@ -85,8 +88,6 @@ app.get('/login', passport.authenticate('auth0'), function(req, res, next) {
 
 const sharedsession = require('express-socket.io-session');
 
-app.use(session); // ATTACH SESSION
-// console.log("initial", session) //Session exists at this point
 io.use(
   //SHARE SESSION WITH IO SOCKETS
   sharedsession(session, {
@@ -100,18 +101,19 @@ app.use(json());
 app.use(cors());
 
 massive(connectionString)
-  .then(dbInstance => app.set('db', dbInstance) && dbInstance.log_them_out())
-  .catch(console.log);
+  .then(dbInstance => app.set("db", dbInstance) & dbInstance.log_them_out())
+  .catch(console.log)
 
 //SOCKET.IO STARTS
 
-let interval;
-io.sockets.on('connection', socket => {
-  var db = app.get('db');
-  socket.handshake.session.user = you;
-  you = null;
-  console.log(you);
-  console.log('Client connected!');
+let interval
+
+io.sockets.on("connection", socket => {
+  var db = app.get("db")
+  socket.handshake.session.user = you
+  you = null
+  console.log(you)
+  console.log("Client connected!")
 
   socket.handshake.session.user
     ? db.run(
@@ -124,36 +126,33 @@ io.sockets.on('connection', socket => {
   var intervalId = setInterval(
     () => getInfoAndEmit(socket, socket.handshake.session.user),
     5000
-  );
-  socket.on('disconnect', () => {
-    console.log('Client disconnected!');
-    clearInterval(intervalId);
+  )
+  socket.on("disconnect", () => {
+    console.log("Client disconnected!")
+    clearInterval(intervalId)
 
     socket.handshake.session.user
       ? db.run(
-          `UPDATE users SET logged_in = false WHERE user_id =${
+          `UPDATE users SET logged_in = false, paired = null WHERE user_id =${
             socket.handshake.session.user.user_id
-          }`
-        ) &
-        db.run(`UPDATE users
-    set waiting_type = 'none'
-    where user_id = ${socket.handshake.session.user.user_id}`) &
-        db.run(
-          `DELETE FROM questions WHERE user_id = ${
+          }; UPDATE users
+          set waiting_type = 'none'
+          where user_id = ${socket.handshake.session.user.user_id}; DELETE FROM questions WHERE user_id = ${
             socket.handshake.session.user.user_id
-          } AND answered = true `
-        ) &
+          } AND (answered = true OR topic_id is NULL)`) &
         console.log(socket.handshake.session.user.user_id) &
         delete socket.handshake.session.user &
         breakSession()
       : // console.log("this one is after the delete",socket.handshake.session.user.user_id)
-        null;
-  });
-});
+        null
+  })
+})
 
 const breakSession = function() {
-  if (session.user.rank === 3) {
-    delete session.user;
+  if (session.user) {
+    if (session.user.rank === 3) {
+      delete session.user
+    }
   }
 };
 
@@ -172,10 +171,14 @@ const getInfoAndEmit = async (socket, usr) => {
       } AND campus_id = ${usr.campus_id} OR waiting_type = 'helping'`
     )
     const res = await db.run(
-      `SELECT * FROM questions WHERE cohort_id = ${
-        usr.cohort_id
-      } AND campus_id = ${usr.campus_id}`
-    );
+      `SELECT users.user_id, users.name, users.image_url, users.paired, questions.q_id, questions.question, questions.code_block, questions.answered, questions.time, topics.topic, topics.color, questions.q_id 
+      FROM users
+        FULL JOIN questions ON users.user_id=questions.user_id 
+        FULL JOIN topics ON topics.id=questions.topic_id 
+        WHERE answered = false AND users.cohort_id = ${
+          usr.cohort_id
+        } AND users.campus_id = ${usr.campus_id} ORDER BY questions.q_id DESC`
+    )
     const currentUser = await db.run(
       `SELECT * FROM users WHERE user_id = ${usr.user_id}`
     );
@@ -253,13 +256,13 @@ app.post('/api/searchSpecificQuestions', controller.getSpecificQuestions);
 //USER CHANGES VIEWS - REMOVES ACTIVE QUESTIONS
 app.put('/api/removequestions', controller.helpRemover);
 
-app.get('/api/me', function(req, res) {
-  console.log(session.user);
-  if (!session.user) {
-    return res.status(404).send('no_user');
+app.get("/api/me", function(req, res) {
+  console.log("from api me ", req.session.user)
+  if (!req.session.user) {
+    return res.status(403).send("no_user")
   }
-  res.status(200).json(session.user);
-  res.status(408).send('no_user');
-});
+  res.status(200).json(req.session.user)
+  res.status(408).send("no_user")
+})
 
-server.listen(port, () => console.log(`Listening on port ${port}`));
+server.listen(port, () => console.log(`Listening on port ${port}`))
